+++ src/ts/Ads/Views/VideoOverlay.ts
@@ -54,6 +54,7 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
     protected _callButtonElement: HTMLElement;
     private _timerElement: HTMLElement;
     private _chinaAdvertisementElement: HTMLElement;
+    private _timerButton: HTMLElement;
 
     private _fadeTimer?: number;
     private _areControlsVisible: boolean = false;
@@ -63,6 +64,8 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
     private _campaign: Campaign;
 
     private _useCloseIconInsteadOfSkipIcon: boolean | undefined = false;
+
+    private _skipUnderTimerExperimentEnabled: boolean = false;
 
     constructor(parameters: IVideoOverlayParameters<Campaign>, privacy: AbstractPrivacy, showGDPRBanner: boolean, showPrivacyDuringVideo: boolean) {
         super(parameters.platform, 'video-overlay', parameters.placement.muteVideo());
@@ -84,6 +87,11 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
         if (this._campaign instanceof PerformanceCampaign || this._campaign instanceof XPromoCampaign) {
             this._templateData.showInstallButton = true;
             this._templateData.gameIcon = this._campaign.getGameIcon() ? this._campaign.getGameIcon().getUrl() : '';
+        }
+
+        if (CustomFeatures.isSkipUnderTimerExperimentEnabled(parameters.coreConfig, parameters.placement)) {
+            this._templateData._skipUnderTimerExperimentEnabled = true;
+            this._skipUnderTimerExperimentEnabled = true;
         }
 
         this._bindings = [
@@ -200,13 +208,22 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
         this._videoProgress = value;
         this._skipRemaining = this._skipDuration - this._videoProgress;
 
-        const timerCount = Math.ceil((this._videoDuration - this._videoProgress) / 1000);
+        let timerCount: number;
+
+        if (this._skipUnderTimerExperimentEnabled) {
+            timerCount = Math.ceil((this._skipRemaining) / 1000);
+        } else {
+            timerCount = Math.ceil((this._videoDuration - this._videoProgress) / 1000);
+        }
 
         if (typeof timerCount === 'number' && !isNaN(timerCount) && timerCount > 0) {
             this._timerElement.innerText = timerCount.toString();
         }
 
         if (this._skipRemaining <= 0) {
+            if (this._skipUnderTimerExperimentEnabled) {
+                this.hideTimerButton();
+            }
             this.showSkipButton();
             this._chinaAdvertisementElement.classList.add('with-skip-button');
         }
@@ -373,6 +390,7 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
         this._callButtonElement = <HTMLElement>this._container.querySelector('.call-button');
         this._timerElement = <HTMLElement>this._container.querySelector('.timer');
         this._chinaAdvertisementElement = <HTMLLIElement>this._container.querySelector('.china-advertisement');
+        this._timerButton = <HTMLElement>this._container.querySelector('.timer-button');
     }
 
     private showSkipButton() {
@@ -382,6 +400,10 @@ export class VideoOverlay extends AbstractVideoOverlay implements IPrivacyHandle
                 this.showCallButton();
             }
         }
+    }
+
+    private hideTimerButton() {
+        this._timerButton.style.display = 'none';
     }
 
     private resetFadeTimer() {